\name{dynsbmmeta-package}
\alias{dynsbmmeta-package}
\alias{dynsbmmeta}
\docType{package}
\title{\packageTitle{dynsbmmeta}}
\description{\packageDescription{dynsbmmeta}}
\section{Package Content}{\packageIndices{dynsbmmeta}}
\author{\packageAuthor{dynsbmmeta}}
\section{Maintainer}{\packageMaintainer{dynsbmmeta}}
\keyword{package}
